Jesus christ.
Version:        2.0.0
Author(s):      Anthony Pizzimenti
Date:           March 18, 2018

This took SO long to get done. Finding the bug that prevented correct
logouts was nigh impossible, and it turns out that my requests to
/api/logout were getting denied with CORS errors because an
empty/deleted href property on a link *still reloads the page when
clicked*. I was trying to send a request while the page was in between
loads, and I nearly screwed myself.

Anyway, I figured it out and everything is fine now.