Refactor to take advantage of new onLoad() flow.
Version:        2.0.4
Author(s):      Anthony Pizzimenti
Date:           April 20, 2018

Refactored nearly all the scripts to get on the new onLoad() bandwagon.
Everything should be working properly now.

Also fixed a small bug on the editor page. Previously, when a user
edited a new post and tried to save it, the 'published' checkbox would
be grabbed, but not passed to saveNewPost(), resulting in a
ReferenceError.